fix: add Object.values polyfill for node <=6
